Switch pdb file format to DebugFull and add pdb files for static libs
The DebugFull format is the only one supported by Dr. Memory. Which
doesn't matter a lot for our dependencies (still nice to have when
we have an error in json_free), but more useful for the thcrap
binaries, and we might as well use it everywhere. I mean, it has
"Full" in its name, it has to be better, right?

And I never noticed those pdb files in the obj directory for static
libraries. Hopefully this is how you're supposed to include them and
it will finally work on other peoples' machines. And hopefully the
.lib files don't contain an absolute path to these files.