Update all dependencies
drlibs changed their API and I can't be bothered to update our code to use
the new one, so we're staying on the previous version for that one.

Switched curl from building with a vcxproj project to building with cmake.
This took a while, but will hopefully make future updates easier.